Trait eosio_cdt::TableCursor

pub trait TableCursor<T>: IntoIterator where
    T: Table, {
    fn bytes(&self) -> Vec<u8>;

    fn erase(&self) -> Result<T::Row, ReadError>;

    fn modify<I: Borrow<T::Row>>(
        &self, 
        payer: Payer, 
        item: I
    ) -> Result<usize, WriteError>;

    fn stream(&self) -> DataStream { ... }

    fn get(&self) -> Result<T::Row, ReadError> { ... }
}

Table Cursor

Required methods

fn bytes(&self) -> Vec<u8>

fn erase(&self) -> Result<T::Row, ReadError>

Erase the current row

Errors

Will return Err if there was an issue reading the stored value. Stored values must be read in order to erase secondary indexes.

fn modify<I: Borrow<T::Row>>(
&self,
payer: Payer,
item: I
) -> Result<usize, WriteError>

Modify the current row

Errors

Will return Err if there was an issue serializing the value.

Loading content...
Provided methods

fn stream(&self) -> DataStream

fn get(&self) -> Result<T::Row, ReadError>

Read and deserialize the current table row

Errors

Will return Err if there was an issue reading the stored value.

Loading content...